Refactor: compute all stats in one place
