--- src/Utilities/Math.js
@@ -1,5 +1,0 @@
-export const average = (...numbers) => {
-    if (numbers.length === 0) return 0;
-    const sum = numbers.reduce((acc, num) => acc + num, 0);
-    return sum / numbers.length;
-};
